mesh_com: Mesh Node split (node/executor)
