+++ modules/mesh_com/mesh_com/mesh_subscriber.py
@@ -3,27 +3,10 @@ from rclpy.node import Node
 from rclpy.qos import QoSPresetProfiles
 
 from std_msgs.msg import String
-
-import json
-import subprocess
-from shlex import quote
+import socket
 
 
 class MeshSubscriber(Node):
-    class MeshSettings:
-        def __init__(self):
-            self.api_version = 1
-            self.ssid = ""
-            self.key = ""
-            self.ap_mac = ""
-            self.country = ""
-            self.frequency = ""
-            self.ip = ""
-            self.subnet = ""
-            self.tx_power = ""
-            self.mode = ""
-            # self.enc = ""
-
     def __init__(self):
         super().__init__('mesh_subscriber')
         self.subscription = self.create_subscription(
@@ -32,61 +15,42 @@ class MeshSubscriber(Node):
             self.listener_callback,
             QoSPresetProfiles.SYSTEM_DEFAULT.value)
         self.subscription  # prevent unused variable warning
-        self.settings = self.MeshSettings()
+        self.HOST = '127.0.0.1'  # Standard loopback interface address (localhost)
+        self.PORT = 33221  # Port to listen on (non-privileged ports are > 1023)
+        self.mesh_socket = 0
+
+        self.backup_timer = 0
+        self.backup_data = 0
+
+    def setup_socket(self):
+        self.mesh_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        self.mesh_socket.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+        self.mesh_socket.settimeout(1)
+
+        self.mesh_socket.connect((self.HOST, self.PORT))
 
     def listener_callback(self, msg):
-        self.get_logger().info('mesh subscriber: "%s"' % msg.data)
-        if msg.data:
-            self.__handle_msg(msg.data)
-
-    def __handle_msg(self, msg):
-
-        # {
-        #     "api_version": 1,                 interface version for future purposes
-        #     "ssid": "gold",                   0-32 octets, UTF-8, shlex.quote chars limiting
-        #     "key": "foobar",                  key for the network
-        #     "ap_mac": "00:11:22:33:44:55",    bssid for mesh network
-        #     "country": "UA",                  Country code, sets tx power limits and supported
-        #                                       channels
-        #     "frequency": "5220",              wifi channel frequency, depends on the country
-        #                                       code and HW
-        #     "ip": "192.168.1.1",              select unique IP
-        #     "subnet": "255.255.255.0",
-        #     "tx_power": "30",                 select 30dBm, HW and regulations limiting it
-        #                                       correct level.
-        #                                       Can be used to set lower dBm levels for testing
-        #                                       purposes (e.g. 5dBm)
-        #     "mode": "mesh"                    mesh=mesh network, ap=debug hotspot
-        # }
-
         try:
-            parameters = json.loads(msg)
-            self.settings.ssid = parameters["ssid"]
-            self.settings.key = parameters["key"]
-            self.settings.ap_mac = parameters["ap_mac"]
-            self.settings.country = parameters["country"].lower()
-            self.settings.frequency = parameters["frequency"]
-            self.settings.ip = parameters["ip"]
-            self.settings.subnet = parameters["subnet"]
-            self.settings.tx_power = parameters["tx_power"]
-            self.settings.mode = parameters["mode"]
-            # self.settings.enc = parameters["enc"]
-            self.__change_configuration()
-        except json.decoder.JSONDecodeError or KeyError or Exception:
-            self.get_logger().info('Setting Failed')
-            pass
+            self.get_logger().info('mesh subscriber: "%s"' % msg.data)
+            if msg.data:
+                self.setup_socket()
+                self.mesh_socket.sendall(str.encode(msg.data))
+                self.mesh_socket.close()
+        except:
+            self.backup_data = msg.data
+            self.backup_timer = self.create_timer(
+                1.0, self.backup_caller)
 
-    def __change_configuration(self):
-        subprocess.call(["/opt/ros/foxy/share/bin/mesh-ibss.sh", quote(self.settings.mode),
-                         quote(self.settings.ip),
-                         quote(self.settings.subnet),
-                         quote(self.settings.ap_mac),
-                         quote(self.settings.key),
-                         quote(self.settings.ssid),
-                         quote(self.settings.frequency),
-                         quote(self.settings.tx_power),
-                         quote(self.settings.country)])
-        self.get_logger().info('Setting Done')
+    def backup_caller(self):
+        try:
+            self.backup_timer.destroy_timer()
+            self.get_logger().info('backup_caller: "%s"' % self.backup_data)
+            self.setup_socket()
+            self.mesh_socket.sendall(str.encode(self.backup_data))
+            self.mesh_socket.close()
+        except:
+            self.backup_timer = self.create_timer(
+                1.0, self.backup_caller)
 
 
 def main(args=None):
